sanitize redirect
