+++ server/src/@types/DecodedPath.d.ts
@@ -1,0 +1,4 @@
+export type DecodedPath = {
+    relative: string;
+    media: string;
+}
