+++ server/src/middleware/route.middleware.ts
@@ -3,36 +3,32 @@ import path from "node:path";
 import { NextFunction, Request, Response } from "express";
 
 import { FileResponse } from "../@types/FileResponse";
+import { DecodedPath } from "../@types/DecodedPath";
 import * as fsService from "../services/fs.service";
 import * as videoService from "../services/video.service";
 import * as subtitleService from "../services/subtitle.service";
 import * as renderService from "../services/render.service";
 import { sanitizeStatict, sanitizeVideo } from "../services/sanitizer.service";
 
-type DecodedPaths = {
-    relative: string;
-    media: string;
-}
-
 // TODO: req.* sanitizer service
 
 export async function routeMiddleware(req: Request, res: Response, next: NextFunction) {
-    const decodedPaths = decodePaths(req.path);
+    const decodedPath = fsService.decodePath(req.path);
 
-    const isDir = await fsService.dirExists(decodedPaths.media);
+    const isDir = await fsService.dirExists(decodedPath.media);
     if (isDir)
-        return dirIndexResponse(decodedPaths, res, next);
+        return dirIndexResponse(decodedPath, res, next);
 
-    const isFile = await fsService.fileExists(decodedPaths.media);
-    const fileExtension = path.extname(decodedPaths.relative).toLowerCase();
+    const isFile = await fsService.fileExists(decodedPath.media);
+    const fileExtension = path.extname(decodedPath.relative).toLowerCase();
 
     let fileHandler: (() => Promise<FileResponse | undefined>) | undefined = undefined;
     switch (true) {
         case fileExtension === ".mp4" && isFile && !sanitizeStatict(req.params["static"]):
-            fileHandler = () => videoService.viewData(decodedPaths.media, decodedPaths.relative, fileExtension);
+            fileHandler = () => videoService.viewData(decodedPath.media, decodedPath.relative, fileExtension);
             break;
         case fileExtension === ".vtt" && !sanitizeStatict(req.params["static"]):
-            fileHandler = () => subtitleService.viewData(decodedPaths.media, sanitizeVideo(req.params["video"]));
+            fileHandler = () => subtitleService.viewData(decodedPath.media, sanitizeVideo(req.params["video"]));
             break;
         default:
             break;
@@ -45,7 +41,7 @@ export async function routeMiddleware(req: Request, res: Response, next: NextFun
     }
 
     if (isFile)
-        return res.sendFile(decodedPaths.media, {
+        return res.sendFile(decodedPath.media, {
             root: process.cwd(),
             dotfiles: "allow"
         });
@@ -53,7 +49,7 @@ export async function routeMiddleware(req: Request, res: Response, next: NextFun
     return next();
 }
 
-function dirIndexResponse(decodedPaths: DecodedPaths, res: Response, next: NextFunction) {
+function dirIndexResponse(decodedPaths: DecodedPath, res: Response, next: NextFunction) {
     return Promise.resolve()
         .then(() => fsService.readDir(decodedPaths.relative, decodedPaths.media))
         .then(items => renderService.renderPage("dir-index", {
@@ -62,14 +58,4 @@ function dirIndexResponse(decodedPaths: DecodedPaths, res: Response, next: NextF
         }))
         .then(({ mime, data }) => res.setHeader("Content-type", mime).end(data))
         .catch(err => next(err));
-}
-
-function decodePaths(requestPath: string): DecodedPaths {
-    const relative = /^[\\/]$/g.test(requestPath)
-        ? ""
-        : decodeURIComponent(requestPath);
-    const media = path.normalize(relative === ""
-        ? "media"
-        : path.join("media", relative));
-    return { relative, media };
 }
